split filters out of search component
